Add files via upload
FIJI Macro which allows user to select one or more intra-nuclear regions from a cell nuclei, and then computes mean gray value of the selected region for all channels, all frames. The user runs the script, selects a region, checks that the region is within the nucleus in all time-frames, and then presses "OK". The gray values will be recorded for each channel and each frame. Each new region is appended to the previous one. The results are in the Measurements window of FIJI and can be saved manually as a .csv file.